git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@11411 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/segment.h
@@ -285,7 +285,7 @@ class CSide {
 
 		inline ubyte Shape (void) { return m_nShape; }
 		inline ubyte FaceCount (void) { return m_nFaces; }
-		void RemapVertices (ushort* map);
+		void RemapVertices (byte* map);
 
 	private:
 		void SetupCorners (ushort* verts, ushort* index);
