+++ main/side.cpp
@@ -980,7 +980,7 @@ return vertices;
 //------------------------------------------------------------------------------
 // Move all vertex id indices > nDeletedIndex down one step
 
-void CSide::RemapVertices (ushort* map)
+void CSide::RemapVertices (byte* map)
 {
 for (int i = 0; i < m_nCorners; i++)
 	m_corners [i] = m_corners [map [i]];
